BatteryStatusChip: Use a lighter background when using light mode
This partially reverts commit 96393e12dfa98411b3511dff2ff65a69f4fde469.

Commit 3cfaf6c132d9e ("SystemUI: Follow light/dark theme in quick
settings") changed the background color of the QS tiles to be darker
when using light theme so that the QS tiles stand out and don't blend
in with the background. However, this is not necessary for the battery
status chip, so parsing the background color for it from the light QS
tiles makes the chip background dark for no reason. Revert it to make
it parse the colors for the dark QS, which makes the chip background
light again.

Change-Id: I67539652714d3f4943f34cdf43b124849281a4db
Signed-off-by: Iverz <asterixiverz@gmail.com>